pyproject.toml
[project]
name = "gemini-core"
version = "0.1.0"
description = ""
authors = [
    {name = "Benjamin Ballintyn",email = "benballintyn@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "google-genai>=1.52.0,<2.0.0",
    "pydantic>=2.12.4,<3.0.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.1,<2.0.0",
    "pandas>=2.3.3,<3.0.0"
]

[tool.poetry.group.local.dependencies]
# This tells Poetry: "When installing this project, link to this folder"
cogito = { path = "../cogito", develop = true }

[tool.poetry.group.dev.dependencies]
ipython = "^8.23.0"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
coverage = "^7.6.10"
codecov = "^2.1.13"
pytest-asyncio = "^0.25.2"
pytest-mock = "^3.14.0"
sphinx = "^8.1.3"
sphinx-autodoc-typehints = "^3.0.1"
sphinx-rtd-theme = "^3.0.2"
mkdocs = "^1.6.1"
mkdocs-material = "^9.7.0"
mkdocstrings = {extras = ["python"], version = "^0.30.1"}

[tool.poetry]
packages = [{include = "gemini_core", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
